+++ src/ResourceManager/Sql/Commands.Sql.Test/ScenarioTests/ManagedInstanceCrudScenarioTests.ps1
@@ -31,7 +31,8 @@ function Test-CreateManagedInstance
  	$licenseType = "BasePrice"
   	$storageSizeInGB = 32
  	$vCore = 16
- 	$skuName = "GP_Gen4"
+	$skuName = "GP_Gen4"
+	$collation = "Serbian_Cyrillic_100_CS_AS"
 
  	try
  	{
@@ -65,7 +66,8 @@ function Test-CreateManagedInstance
 		# With edition and computeGeneration specified
  		$job = New-AzureRmSqlInstance -ResourceGroupName $rg.ResourceGroupName -Name $managedInstanceName `
  			-Location $rg.Location -AdministratorCredential $credentials -SubnetId $subnetId `
-  			-LicenseType $licenseType -StorageSizeInGB $storageSizeInGB -Vcore $vCore -Edition $edition -ComputeGeneration $computeGeneration -DnsZonePartner $dnsZonePartner -AsJob
+			-LicenseType $licenseType -StorageSizeInGB $storageSizeInGB -Vcore $vCore -Edition $edition `
+			-ComputeGeneration $computeGeneration -DnsZonePartner $dnsZonePartner -Collation $collation -AsJob
  		$job | Wait-Job
  		$managedInstance1 = $job.Output
 
@@ -78,6 +80,7 @@ function Test-CreateManagedInstance
 		Assert-AreEqual $managedInstance1.LicenseType $licenseType
 		Assert-AreEqual $managedInstance1.VCores $vCore
 		Assert-AreEqual $managedInstance1.StorageSizeInGB $storageSizeInGB
+		Assert-AreEqual $managedInstance1.Collation $collation
  		Assert-StartsWith ($managedInstance1.ManagedInstanceName + ".") $managedInstance1.FullyQualifiedDomainName
  	}
  	finally