+++ src/ResourceManager/Sql/Commands.Sql/ManagedInstance/Cmdlet/NewAzureSqlManagedInstance.cs
@@ -147,6 +147,15 @@ namespace Microsoft.Azure.Commands.Sql.ManagedInstance.Cmdlet
         public string DnsZonePartner { get; set; }
 
         /// <summary>
+        /// Gets or sets the instance collation
+        /// </summary>
+        [Parameter(Mandatory = false,
+            HelpMessage = "The collation of the Azure SQL Managed Instance to use.")]
+        [ValidateNotNullOrEmpty]
+        [PSArgumentCompleter("SQL_Latin1_General_CP1_CI_AS", "Latin1_General_100_CS_AS_SC")]
+        public string Collation { get; set; }
+
+        /// <summary>
         /// Gets or sets the tags to associate with the instance
         /// </summary>
         [Parameter(Mandatory = false,
@@ -237,7 +246,8 @@ namespace Microsoft.Azure.Commands.Sql.ManagedInstance.Cmdlet
                 SubnetId = this.SubnetId,
                 VCores = this.VCore,
                 Sku = Sku,
-                DnsZonePartner = this.DnsZonePartner
+                DnsZonePartner = this.DnsZonePartner,
+                Collation = this.Collation
             });
             return newEntity;
         }