+++ src/ResourceManager/Sql/Commands.Sql/ManagedInstance/Model/AzureSqlManagedInstanceModel.cs
@@ -107,5 +107,10 @@ namespace Microsoft.Azure.Commands.Sql.ManagedInstance.Model
         /// Gets or sets the dns zone of the managed instance
         /// </summary>
         public string DnsZone { get; set; }
+
+        /// <summary>
+        /// Gets or sets the Managed Instance collation
+        /// </summary>
+        public string Collation { get; set; }
     }
 }